requirements.txt
python-telegram-bot==13.7
telegram==0.0.1
requests==2.31.0
base58==2.1.1
bech32==1.2.0
ecdsa==0.19.0
#bitcoinlib
#aiohttp
requests==2.31.0
blockcypher==1.0.93
hdwallet==2.2.1
solana==0.34.3
solders==0.21.0
solathon==1.0.2
#mysql-connector-python==9.0.0
httpx==0.27.0
sqlalchemy
pymysql
eth_account
python-dotenv==1.0.1
cryptography==43.0.0
psutil==6.0.0
#Python 3.8.10
#Ah this annoying Solana and solathon, KYS FRR